bin/ci/structured_logger.py
   """
   A structured logging utility supporting multiple data formats such as CSV, JSON,
   and YAML.
      The main purpose of this script, besides having relevant information available
   in a condensed and deserialized.
      This script defines a protocol for different file handling strategies and provides
   implementations for CSV, JSON, and YAML formats. The main class, StructuredLogger,
   allows for easy interaction with log data, enabling users to load, save, increment,
   set, and append fields in the log. The script also includes context managers for
   file locking and editing log data to ensure data integrity and avoid race conditions.
   """
      import json
   import os
   from collections.abc import MutableMapping, MutableSequence
   from contextlib import contextmanager
   from datetime import datetime
   from pathlib import Path
   from typing import Any, Protocol
      import fire
   from filelock import FileLock
      try:
                  except ImportError as e:
            try:
                  except ImportError as e:
               class ContainerProxy:
      """
   A proxy class that wraps a mutable container object (such as a dictionary or
   a list) and calls a provided save_callback function whenever the container
   or its contents
   are changed.
   """
   def __init__(self, container, save_callback):
      self.container = container
         def __getitem__(self, key):
      value = self.container[key]
   if isinstance(value, (MutableMapping, MutableSequence)):
               def __setitem__(self, key, value):
      self.container[key] = value
         def __delitem__(self, key):
      del self.container[key]
         def __getattr__(self, name):
               if callable(attr):
         def wrapper(*args, **kwargs):
                              def __iter__(self):
            def __len__(self):
            def __repr__(self):
            class AutoSaveDict(dict):
      """
   A subclass of the built-in dict class with additional functionality to
   automatically save changes to the dictionary. It maintains a timestamp of
   the last modification and automatically wraps nested mutable containers
   using ContainerProxy.
   """
            def __init__(self, *args, save_callback, register_timestamp=True, **kwargs):
      self.save_callback = save_callback
   self.__register_timestamp = register_timestamp
   self.__heartbeat()
   super().__init__(*args, **kwargs)
         def __heartbeat(self):
      if self.__register_timestamp:
         def __save(self):
      self.__heartbeat()
         def __wrap_dictionaries(self):
      for key, value in self.items():
         if isinstance(value, MutableMapping) and not isinstance(
         ):
               def __setitem__(self, key, value):
      if isinstance(value, MutableMapping) and not isinstance(value, AutoSaveDict):
         value = AutoSaveDict(
                  if self.__register_timestamp and key == AutoSaveDict.timestamp_key:
               def __getitem__(self, key):
      value = super().__getitem__(key)
   if isinstance(value, (MutableMapping, MutableSequence)):
               def __delitem__(self, key):
      super().__delitem__(key)
         def pop(self, *args, **kwargs):
      result = super().pop(*args, **kwargs)
   self.__save()
         def update(self, *args, **kwargs):
      super().update(*args, **kwargs)
   self.__wrap_dictionaries()
         class StructuredLoggerStrategy(Protocol):
      def load_data(self, file_path: Path) -> dict:
            def save_data(self, file_path: Path, data: dict) -> None:
            class CSVStrategy:
      def __init__(self) -> None:
      if CSV_LIB_EXCEPTION:
         raise RuntimeError(
         def load_data(self, file_path: Path) -> dict:
      dicts: list[dict[str, Any]] = pl.read_csv(
         ).to_dicts()
   data = {}
   for d in dicts:
         for k, v in d.items():
      if k != AutoSaveDict.timestamp_key and k in data:
      if isinstance(data[k], list):
         data[k].append(v)
               def save_data(self, file_path: Path, data: dict) -> None:
            class JSONStrategy:
      def load_data(self, file_path: Path) -> dict:
            def save_data(self, file_path: Path, data: dict) -> None:
      with open(file_path, "w") as f:
         class YAMLStrategy:
      def __init__(self):
      if YAML_LIB_EXCEPTION:
         raise RuntimeError(
         self.yaml = YAML()
   self.yaml.indent(sequence=4, offset=2)
   self.yaml.default_flow_style = False
         @classmethod
   def represent_dict(cls, dumper, data):
            def load_data(self, file_path: Path) -> dict:
            def save_data(self, file_path: Path, data: dict) -> None:
      with open(file_path, "w") as f:
         class StructuredLogger:
      def __init__(
         ):
      self.file_name: str = file_name
   self.file_path = Path(self.file_name)
            if strategy is None:
         self.strategy: StructuredLoggerStrategy = self.guess_strategy_from_file(
         else:
            if not self.file_path.exists():
         Path.mkdir(self.file_path.parent, exist_ok=True)
            if truncate:
         with self.get_lock():
         def load_data(self):
            def save_data(self):
            @property
   def data(self) -> AutoSaveDict:
            @contextmanager
   def get_lock(self):
      with FileLock(f"{self.file_path}.lock", timeout=10):
         @contextmanager
   def edit_context(self):
      """
   Context manager that ensures proper loading and saving of log data when
   performing multiple modifications.
   """
   with self.get_lock():
         try:
      self.load_data()
            @staticmethod
   def guess_strategy_from_file(file_path: Path) -> StructuredLoggerStrategy:
      file_extension = file_path.suffix.lower().lstrip(".")
         @staticmethod
   def get_strategy(strategy_name: str) -> StructuredLoggerStrategy:
      strategies = {
         "csv": CSVStrategy,
   "json": JSONStrategy,
   "yaml": YAMLStrategy,
            try:
         except KeyError as e:
         if __name__ == "__main__":
      fire.Fire(StructuredLogger)
